User-spawned particles can now have an initial velocity
The user is able to spawn a particle with an initial velocity. He can do this by right-clicking at the place where he wants the particle to spawn, and dragging his cursor across the screen to set the direction and velocity of the particle. Finally, the user can release the right mouse button to spawn the particle, which will travel in the direction which the user has set.

When the user drags the mouse for a longer distance, the particle will have a faster speed.