Improved Physics Engine
Some bugs were removed. Previously, when a particle was placed inside another particle, the particles would merge, and the newly-created particle will zoom off with a very high speed. This was fixed by making sure that the forces of a particle are only calculated if it is not colliding with another particle.